TICKET-141: Split PlatformNode into focused modules
Extract three focused modules from the 825-line PlatformNode:
- platform/textures.ts: 5 texture generator functions
- platform/wake.ts: wake trail tracking, rasterizeWake, WakeTrailPoint
- platform/shaderPatch.ts: ripple/wake/hit-ripple shader patching

PlatformNode becomes a concise orchestrator that composes these pieces.
All public exports are re-exported for backward compatibility.
